+++ src/compose.ts
@@ -1,15 +1,14 @@
 import type { Last } from './utilities';
 
 /** Performs right-to-left function composition. */
-export function compose<T extends Array<(...args: ReadonlyArray<any>) => any>>(
-  ...fns: T
-): (...args: Parameters<Last<T>>) => ReturnType<T[0]> {
+export function compose<
+  T extends ReadonlyArray<(...args: ReadonlyArray<any>) => any>
+>(...fns: T): (...args: Parameters<Last<T>>) => ReturnType<T[0]> {
   return (...args) => {
-    const firstFn = fns.pop() ?? ((...x) => x);
+    const firstFn = fns[fns.length - 1] ?? ((...x) => x);
 
-    return fns.reduceRight(
-      (currentArgs, fn) => fn(currentArgs),
-      firstFn(...args)
-    );
+    return fns
+      .slice(0, fns.length - 1)
+      .reduceRight((currentArgs, fn) => fn(currentArgs), firstFn(...args));
   };
 }